fix: move imports to top level, remove docs/plans
Address PR review feedback on import placement:

- Extract _handle_unexpected_error to _errors.py to break circular
  imports between main.py and CLI subcommand modules
- Move lazy imports to top level in _tailor.py, _screen.py
- Use top-level try/except for optional litellm in client.py, gate.py
- Move lazy imports to top level in resume_extensions.py, loader.py
- Keep linkedin imports lazy in _import.py (optional [linkedin] dep)
- Remove docs/plans/ design doc (content captured in PR description)
- Fix missing import re in tailor.py _check_skills
- Update test mock paths to match new import locations
