+++ src/simple_resume/shell/cli/_import.py
@@ -13,6 +13,7 @@ from pathlib import Path
 import yaml
 
 from simple_resume.core.exceptions import SimpleResumeError
+from simple_resume.shell.cli._errors import _handle_unexpected_error
 
 logger = logging.getLogger(__name__)
 
@@ -28,11 +29,10 @@ def handle_import_command(args: argparse.Namespace) -> int:
         Exit code (0 for success, non-zero for failure).
 
     """
-    from simple_resume.core.importers.linkedin import (  # noqa: PLC0415
+    from simple_resume.core.importers.linkedin import (  # noqa: PLC0415 - optional dep
         linkedin_to_simple_resume,
     )
-    from simple_resume.shell.cli.main import _handle_unexpected_error  # noqa: PLC0415
-    from simple_resume.shell.importers.linkedin_fetcher import (  # noqa: PLC0415
+    from simple_resume.shell.importers.linkedin_fetcher import (  # noqa: PLC0415 - optional dep
         read_linkedin_file,
     )
